Another Implementation of Provider State Management
Used provider to manage the state of a basic messaging app. This app is not connected to any server. It just takes your subject and message and displays it unto another screen. It checks for input validations and provides a floating action button to send the message. If the form is invalid, it will not display the floating button.